Output file name should not change flags picked for compiler (#1938)
Fixes SWDEV-207362,

The output file name should not contribute to picking up the right flags for the compiler. This fix solves issues when the output has conflicting extensions which confuses hipcc to treat them as the source files and add the required flags for them.

PS: Output file refers to the file followed by -o

Change-Id: I8d9e68355e1e162237a2552be425b82e067887e1
Example: hipcc test.o -o test.hip will add the flags for .hip compilation ignoring the fact that it is an output file